removed 8 and 9 as backspace and delete, removed 3-numpad
turns out that backspaces is used too often, extensor of middle finger started to hurt from pressing 8 too often. The middle finger is the heaviest finger, and lifting it up too often is really not the best idea.

Delete is not used so often, so it's ok to not have it near. Might consider , & m in the future

Since 8 and 9 are no longer needed, 3-numpad is no longer needed either.